+++ back-end/api/index.js
@@ -1,3 +1,0 @@
-import { author } from "./author.js";
-
-export { author };