+++ back-end/controller/index.js
@@ -1,0 +1,3 @@
+import { authorController } from "./author.js";
+
+export { authorController };