+++ back-end/Changelog.md
@@ -6,3 +6,4 @@
 
 - Config db configurations and api security
 - Create author,handle validation and exception handling
+- Create get all authors function and implement dto